Feature/issue-53 API Forum Client review fix
