+++ packages/client/src/slices/forumSlice.ts
@@ -1,33 +1,7 @@
 import { createAsyncThunk, createSlice, PayloadAction } from '@reduxjs/toolkit'
 import { RootState } from '@/store'
 import { SERVER_HOST2 } from '@/constants'
-
-export interface Topic {
-  id: number
-  title: string
-  description: string
-  author_id: string
-  createdAt: string
-  updatedAt: string
-}
-
-export interface Comment {
-  id: number
-  content: string
-  author_id: string
-  topic_id: number
-  parentId: number | null
-  createdAt: string
-  updatedAt: string
-  replies?: Comment[]
-}
-
-interface ForumState {
-  topics: Topic[]
-  currentComments: Comment[]
-  isLoading: boolean
-  error: string | null
-}
+import { Topic, Comment, ForumState } from './forum.types'
 
 const initialState: ForumState = {
   topics: [],
